+++ server/src/routes/index.ts
@@ -2,10 +2,18 @@ import assets from './assets'
 import records from './records'
 import advice from './advice'
 import insights from './insights'
+import password from './password'
 import { Assets } from './../models/assets'
 import { Record } from './../models/records'
 
 export const reset = async (_, reply) => {
+  const canBeReset = process.env.CAN_BE_RESET || false
+  if (!canBeReset) {
+    return reply.code(400).send({
+      statusCode: 400,
+      message: 'The database can not be reset.',
+    })
+  }
   try {
     await Assets.destroy({ where: {} })
     await Record.destroy({ where: {} })
@@ -37,4 +45,4 @@ const routes = [
   },
 ]
 
-export default [...routes, ...assets, ...records, ...advice, ...insights]
+export default [...routes, ...assets, ...records, ...advice, ...insights, ...password]
